requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
jinja2>=3.1.4
python-multipart>=0.0.12
authlib>=1.3.2
httpx>=0.27.2
itsdangerous>=2.2.0
bleach>=6.1.0
mistralai>=1.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9
weasyprint>=62.0
Pillow>=11.0.0
pillow-avif-plugin>=1.4.6
